do not sleep forever
there aren't many good reasons to do this, it was initially written
w/ the assumption that in certain failure modes we can just keep
writing the same info out while we wait to presumably be restarted.
in practice there are other types of errors that can happen which should
just get retried as they are probably transient errors.