Restart k3s service when wg tunnel changes (#11)
* Restart k3s service when wg tunnel changes

We have configured the k3s service to bind to the wireguard tunnel
address. However, the IP address of the tunnel may change, and k3s
does not restart automatically in this case, so we implement via
the coordiantor.

TODO: Changing the node IP will cause issues with pods already
scheduled to it. Best practice is to delete and re-register the node
in this case. See:
https://kubernetes.io/docs/concepts/architecture/nodes/#node-name-uniqueness